Before file separation.

+++ temp_api.h
@@ -40,7 +40,7 @@ typedef struct {
 	uint8_t month;
 	uint16_t year;	
 	uint32_t encoded_my;
-	uint16_t count;
+	uint32_t count;
 	int32_t sum_t;
 	int8_t min_t;
 	int8_t max_t;
